# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API server for RaptorChat.",
        "title": "RaptorChat API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@raptorchat.example.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/user/{id}/dm": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves the direct message conversation between the authenticated user and specified recipient",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users",
                    "rooms"
                ],
                "summary": "Get direct message conversation with another user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Recipient ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Direct message object",
                        "schema": {
                            "$ref": "#/definitions/db.Room"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid recipient ID or User ID not found in context",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns a list of all users in the system except the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "List all users (excluding current user)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/orm.User"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized or missing token",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/me": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns the ID of the currently authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get authenticated user's ID",
                "responses": {
                    "200": {
                        "description": "User ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "401": {
                        "description": "Unauthorized or missing token",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/me/activity": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves the authenticated user's 5 most recent messages and calls",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users",
                    "activity"
                ],
                "summary": "Get current user's recent activity",
                "responses": {
                    "200": {
                        "description": "User's recent activity",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "calls": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/orm.Call"
                                    }
                                },
                                "messages": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/db.Message"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "User ID not found in context",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/me/avatar": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Accepts a multipart/form-data file upload for the authenticated user's avatar, stores it, and updates the user's record",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users",
                    "avatars"
                ],
                "summary": "Upload or update user's avatar",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Avatar image file (JPEG/PNG/etc.)",
                        "name": "avatar",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns {\\\"avatar_url\\\": \\\"\u003curl\u003e\\\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad form data or missing file",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized or missing token",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Server error while saving avatar",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Deletes the authenticated user's avatar file and clears the URL in the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users",
                    "avatars"
                ],
                "summary": "Delete user's avatar",
                "responses": {
                    "200": {
                        "description": "Returns {\\\"avatar_url\\\": \\\"\\\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized or missing token",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Server error while deleting avatar",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/me/friends": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns a list of friends (other users) for the currently authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users",
                    "friends"
                ],
                "summary": "Get friends of the authenticated user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/orm.User"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized or missing token",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "db.Room": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "member_count": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "integer"
                },
                "type": {
                    "$ref": "#/definitions/db.RoomsType"
                }
            }
        },
        "db.RoomsType": {
            "type": "string",
            "enum": [
                "direct",
                "group"
            ],
            "x-enum-varnames": [
                "RoomsTypeDirect",
                "RoomsTypeGroup"
            ]
        },
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "orm.Call": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "ended_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "participant_count": {
                    "type": "integer"
                },
                "peak_participant_count": {
                    "type": "integer"
                },
                "room_id": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/orm.CallsStatus"
                }
            }
        },
...
        "orm.User": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}